Rename Fractal helper function file and update imports.
