+++ src/scmultiplex/fractal/surface_mesh.py
@@ -39,7 +39,7 @@ from skimage.measure import label
 from skimage.morphology import disk, remove_small_objects
 from skimage.segmentation import expand_labels
 
-from scmultiplex.fractal.FractalHelperFunctions import get_zattrs, convert_indices_to_origin_zyx, format_roi_table
+from scmultiplex.fractal.fractal_helper_functions import get_zattrs, convert_indices_to_origin_zyx, format_roi_table
 
 from scmultiplex.meshing.FilterFunctions import equivalent_diam, mask_by_parent_object, \
     calculate_mean_volume
